TODos mostly completed FileUtils#readLines method - clarification needed
